added more maintenance items - changed titles, and made rating consistent
