+++ src/bar_plots.py
@@ -1,11 +1,12 @@
 import altair as alt
 import pandas as pd
+from vino_themes import vino_special
 
-x_labels = {'region_1': 'Region',
+X_LABELS = {'region_1': 'Region',
             'winery': 'Winery',
             'variety': 'Grape Variety'}
 
-y_labels = {'points': 'Rating',
+Y_LABELS = {'points': 'Rating',
             'price': 'Price',
             'value_scaled': 'Value'}
 
@@ -22,58 +23,6 @@ def sort_extract_bar_plot(data, y_name='points', x_name='winery', n=15, directio
     direction -- (str) whether to sort the column by ascending or descending
                        order based on the by column (default='desc')
     """
-    def vino_special():
-        font = "Helvetica"
-        axisColor = "#000000"
-        gridColor = "#DEDDDD"
-        return {
-            "config": {
-                "title": {
-                    "fontSize": 24,
-                    "font": font,
-                    "anchor": "start", # equivalent of left-aligned.
-                    "fontColor": "#000000"
-                },
-                'view': {
-                    "height": 300, 
-                    "width": 500
-                },
-                "axisX": {
-                    "domain": True,
-                    #"domainColor": axisColor,
-                    "gridColor": gridColor,
-                    "domainWidth": 1,
-                    "grid": False,
-                    "labelFont": font,
-                    "labelFontSize": 12,
-                    "labelAngle": 60, 
-                    "tickColor": axisColor,
-                    "tickSize": 5, # default, including it just to show you can change it
-                    "titleFont": font,
-                    "titleFontSize": 16,
-                    "titlePadding": 10, # guessing, not specified in styleguide
-                    "title": "X Axis Title (units)", 
-                },
-                "axisY": {
-                    "domain": False,
-                    "grid": True,
-                    "gridColor": gridColor,
-                    "gridWidth": 1,
-                    "labelFont": font,
-                    "labelFontSize": 14,
-                    "labelAngle": 0, 
-                    #"ticks": False, # even if you don't have a "domain" you need to turn these off.
-                    "titleFont": font,
-                    "titleFontSize": 16,
-                    "titlePadding": 10, # guessing, not specified in styleguide
-                    "title": "Y Axis Title (units)", 
-                    # titles are by default vertical left of axis so we need to hack this 
-                    #"titleAngle": 0, # horizontal
-                    #"titleY": -10, # move it up
-                    #"titleX": 18, # move it to the right so it aligns with the labels 
-                },
-            }
-                }
 
     # register the custom theme under a chosen name
     alt.themes.register('vino_special', vino_special)
@@ -94,7 +43,7 @@ def sort_extract_bar_plot(data, y_name='points', x_name='winery', n=15, directio
                       op="sum",  
                       order='descending',  
                   )),
-            alt.Y(y_name + ':Q', title=y_labels[y_name],
+            alt.Y(y_name + ':Q', title=Y_LABELS[y_name],
                   scale=alt.Scale(domain=[min(new_data[y_name]),
                      max(new_data[y_name])])
                  ),
@@ -105,7 +54,7 @@ def sort_extract_bar_plot(data, y_name='points', x_name='winery', n=15, directio
             ),
             tooltip=[alt.Tooltip(f'{x_name}:N'),
                     alt.Tooltip(f'{y_name}:Q')]
-        ).properties(width=500, height=300, title='Average ' + y_labels[y_name] + ' by ' + x_labels[x_name]) 
+        ).properties(width=500, height=300, title='Average ' + Y_LABELS[y_name] + ' by ' + X_LABELS[x_name]) 
         return ranked_bar
     else:
         # If we want to sort from lowest to highest
@@ -117,7 +66,7 @@ def sort_extract_bar_plot(data, y_name='points', x_name='winery', n=15, directio
                       op="sum",  
                       order='ascending'  
                   )),
-            alt.Y(y_name + ':Q', title=y_labels[y_name], 
+            alt.Y(y_name + ':Q', title=Y_LABELS[y_name], 
                   scale=alt.Scale(domain=[min(new_data[y_name])-1,
                              max(new_data[y_name])])
                  ),
@@ -127,5 +76,5 @@ def sort_extract_bar_plot(data, y_name='points', x_name='winery', n=15, directio
                 alt.value('lightgrey')
             ),
             tooltip=[x_name + ':N', y_name + ':Q']
-        ).properties(width=500, height=300, title='Average ' + y_labels[y_name] + ' by ' + x_labels[x_name]) 
+        ).properties(width=500, height=300, title='Average ' + Y_LABELS[y_name] + ' by ' + X_LABELS[x_name]) 
         return ranked_bar
